add basic crud functionality to dbcontext

+++ Models/Entities/CodingTime.cs
@@ -1,3 +1,3 @@
 ﻿namespace Models.Entities;
 
-public record class CodingTime(int Id, string Task, string StartTime, string EndTime);
+public record class CodingTime(System.Int64 Id, string Task, string StartTime, string EndTime);
